added dough function
